refactor: Consolidate filter code into DashboardFiltersContext for improved architecture
